Revert "Performance Improvements & Reader design update"
