--- .babelrc
@@ -1,8 +1,0 @@
-{
-    "presets": [["@babel/preset-typescript", {"isTSX": true, "allExtensions": true}], "@babel/preset-react"],
-    "plugins": [
-        "react-hot-loader/babel",
-        ["@babel/plugin-proposal-class-properties", {"loose": true}],
-        ["babel-plugin-styled-components", {"pure": true}]
-    ]
-}
