--- custom.d.ts
@@ -1,5 +1,0 @@
-declare module '*.svg' {
-    const content: any
-    export const ReactComponent: any
-    export default content
-}
