+++ package.json
@@ -3,50 +3,37 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
-        "@babel/plugin-proposal-class-properties": "^7.5.5",
-        "@babel/preset-typescript": "^7.3.3",
-        "@hot-loader/react-dom": "^16.9.0",
         "@types/jest": "24.0.16",
-        "@types/lodash": "^4.14.136",
         "@types/node": "12.6.8",
         "@types/react": "^16.8.25",
-        "@types/react-dom": "^16.8.5",
+        "@types/react-dom": "16.8.5",
         "@types/react-router-dom": "^4.3.4",
         "@types/react-scroll": "^1.5.4",
         "@types/styled-components": "^4.1.18",
-        "add": "^2.0.6",
         "eslint-config-react-app": "^4.0.1",
         "eslint-plugin-flowtype": "^3.13.0",
         "eslint-plugin-import": "^2.18.2",
-        "fork-ts-checker-webpack-plugin": "^1.5.0",
         "handlebars": "4.1.2",
         "husky": "^2.2.0",
-        "lodash": "^4.17.15",
         "mobx": "^5.9.0",
         "mobx-react": "^6.0.3",
-        "path": "^0.12.7",
         "pretty-quick": "^1.11.1",
         "react": "^16.8.6",
-        "react-dev-utils": "^9.0.3",
         "react-dom": "^16.8.6",
-        "react-hot-loader": "^4.12.11",
         "react-media": "^1.9.2",
         "react-mobx": "^0.0.4",
         "react-router-dom": "^5.0.1",
         "react-scripts": "^2.1.1",
         "react-scroll": "^1.7.11",
         "styled-components": "^4.3.2",
-        "typescript": "3.5.3",
-        "typescript-plugin-styled-components": "^1.4.3",
-        "url-loader": "^2.1.0",
-        "yarn": "^1.17.3"
+        "typescript": "3.5.3"
     },
     "eslintConfig": {
         "extends": "react-app"
     },
     "scripts": {
-        "build": "webpack --mode production --config ./config/webpack.config.js",
-        "start": "node ./config/start.js",
+        "start": "react-scripts start",
+        "build": "react-scripts build",
         "test": "react-scripts test --env=jsdom",
         "eject": "react-scripts eject",
         "typecheck": "tsc --noEmit",
@@ -65,31 +52,15 @@
         ]
     },
     "devDependencies": {
-        "@babel/core": "^7.5.5",
-        "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-        "@babel/plugin-transform-async-to-generator": "^7.5.0",
-        "@babel/preset-env": "^7.5.5",
-        "@babel/preset-react": "^7.0.0",
-        "@svgr/webpack": "^4.3.2",
         "@typescript-eslint/eslint-plugin": "^1.13.0",
         "@typescript-eslint/parser": "^1.13.0",
-        "awesome-typescript-loader": "^5.2.1",
-        "babel-loader": "^8.0.6",
-        "css-loader": "^3.2.0",
         "eslint-config-prettier": "^6.0.0",
         "eslint-plugin-jsx-a11y": "^6.2.3",
         "eslint-plugin-prettier": "^3.1.0",
         "eslint-plugin-react": "^7.14.3",
         "eslint-plugin-react-hooks": "^1.6.1",
-        "html-loader": "^0.5.5",
-        "html-webpack-plugin": "^3.2.0",
         "prettier": "1.18.2",
-        "source-map-loader": "^0.2.4",
-        "style-loader": "^1.0.0",
-        "tslint-config-prettier": "^1.18.0",
-        "webpack": "^4.39.1",
-        "webpack-cli": "^3.3.6",
-        "webpack-dev-server": "^3.8.0"
+        "tslint-config-prettier": "^1.18.0"
     },
     "husky": {
         "hooks": {
