+++ src/Components/Button/Button.css
@@ -1,0 +1,19 @@
+.btn {
+    display:block;
+    margin: auto;
+    width: fit-content;
+    font-family:'Muli', sans-serif;
+    background: #6c63ff;
+    border-radius: 4px;
+    padding: 12px 20px;
+    font-size: 14px;
+    color: #ccf;
+    transition: opacity 0.3s ease-in-out;
+    text-decoration: none;
+    cursor: pointer;
+}
+
+a.btn:hover,
+.btn:hover {
+    opacity: 0.6;
+}
