+++ src/Components/Information/information.css
@@ -1,0 +1,40 @@
+.information {
+    margin: 30px 30px;
+    text-align: left;
+}
+
+.jump {
+    font-family: 'Muli', sans-serif;
+    font-weight: 300;
+    margin: 130px 0 30px 0px;
+    font-size: 25px;
+    color: #101010;
+    letter-spacing: 1.08px;
+}
+
+.back-link {
+    font-size: 0.9em;
+    text-decoration: none;
+    border: none;
+    border-bottom: 1px dotted #fff;
+    padding: 0;
+    margin-bottom: 1.4em;
+}
+
+.link {
+    cursor: pointer;
+}
+
+@media (max-width: 1024px) {
+    .jump {
+        margin: 60px 0 30px 0px;
+    }
+}
+
+li a {
+    text-decoration: none;
+    color: #3c8eff;
+    display: inline-block;
+    border-bottom: 1px dotted;
+} 
+
