+++ src/Components/MainMenu/MainMenu.css
@@ -1,0 +1,46 @@
+.main-nav{
+    line-height: 40px;
+    text-align: center;
+}
+
+.main-nav a {
+    margin-right: 14px;
+    margin-left: 14px;
+    text-decoration: none;
+}
+    .main-nav a > img {        
+        display:none;
+    }
+
+@media (max-width: 780px) {
+
+    .nav-on .main-nav{
+        display: block;
+    }
+
+    .nav-off .main-nav{
+        display: none;
+    }
+    .main-nav{
+        position: fixed;
+        top:0;
+        margin-left: 14px;
+        background: #fff;
+        margin-top:56px;
+        border-radius: 6px;    
+    }
+    .main-nav a {        
+        color: #333333;
+        display: block;
+        margin-right: 0;
+        margin-left: 0;
+        line-height: 40px;
+        padding: 10px 30px;
+    }
+    .main-nav a > img {        
+        display:block;
+        margin: 11px 10px auto 0;
+        height: 1rem;
+        float: left;
+    }
+}
