+++ src/Components/Preferences/Preferences.tsx
@@ -1,9 +1,9 @@
 import React, {useState, useEffect} from 'react'
 import './preferences.css'
-import Overlay from '~views/Overlay/Overlay'
+import Overlay from '../../Views/Overlay/Overlay'
 import DrawerButton from './DrawerButton'
 import ModeButton from './ModeButton'
-import {IEdlizStore} from '~stores/EdlizStore'
+import {IEdlizStore} from '../../stores/EdlizStore'
 
 interface PreferencesProps extends Omit<IEdlizStore, 'navigation' | 'toggleNavigationPanel' | 'books'> {}
