+++ src/Data/Information/acknowledgements.js
@@ -1,0 +1,128 @@
+import React from 'react';
+
+export let acknowledgements = {
+    short_title: "Acknowledgements",
+    long_title: "Acknowledgements",
+    content:
+      <span>
+        <h1>Acknowledgements</h1>
+        <p>We would like to thank all the individuals who made contributions through colleagues or discussion forums or by communicating through electronic mail. We are grateful to all who made this edition a national guide that serves as the standard for Zimbabwe. Thank you to all the healthcare workers for your support</p>
+        <p>The following attended our review workshops as well as being instrumental in current chapter reviews:</p>
+        <ul>
+          <li>Akinjide-Obonyo Akindele P, Dr</li>
+          <li>Maunga Simbarashe, Dr</li>
+          <li>Apollo Tsitsi, Dr</li>
+          <li>Maunganidze Aspect, Dr</li>
+          <li>Bakasa Clemenciana, Ms</li>
+          <li>Mbuzi Tonnie, Mr</li>
+          <li>Bare Blessing, </li>
+          <li>Mhazo Tichatyei, Mr</li>
+          <li>Basopo Victor, Mr</li>
+          <li>Mhembere Josephine, Dr</li>
+          <li>Bepe Tafadzwa, Dr</li>
+          <li>Midzi Stanley, Dr</li>
+          <li>Borok Margaret, Dr</li>
+          <li>Misihairambwi Silence, Ms</li>
+          <li>Burutsa Patricia, Ms</li>
+          <li>Mlilo Lindiwe, Dr</li>
+          <li>Bwakura Tapiwanashe, Dr</li>
+          <li>Moyo Dothan, Mr</li>
+          <li>Cakana Andrew, Prof</li>
+          <li>Moyo Mluleki, Mr</li>
+          <li>Chakanyuka Artmore, Dr</li>
+          <li>Moyo Sifiso, Mr</li>
+          <li>Chakanyuka Christine C., Dr</li>
+          <li>Mudombi Wisdom, Dr</li>
+          <li>Chari Godfrey,</li>
+          <li>Mudzimu Forward, Mr</li>
+          <li>Charimari Lincoln, Dr</li>
+          <li>Mujuru Hilda A., Dr</li>
+          <li>Chemhuru Milton, Dr</li>
+          <li>Mungwadzi Godfrey, Dr</li>
+          <li>Chikanya Sonia Irene, Ms</li>
+          <li>Munjanja Stephen P., Prof</li>
+          <li>Chimhini Gwendoline, Dr</li>
+          <li>Mushavi Angela, Dr</li>
+          <li>Chirenje Mike Z., Prof.</li>
+          <li>Musiya N., Dr</li>
+          <li>Chiro Erick, Mr</li>
+          <li>Musungwa Alexio, Mr</li>
+          <li>Cowan Frances M, Prof.</li>
+          <li>Musvipa Mary, Ms</li>
+          <li>Deda Petunia, Ms</li>
+          <li>Mutsvairo Sitembile, Ms</li>
+          <li>Dliwayo Thokozile, Ms</li>
+          <li>Mwaramba Charles, Mr</li>
+          <li>Dube Siphathisiwe Noreen, Mrs</li>
+          <li>Mwonzora Muchaneta, Mrs</li>
+          <li>Dube Tirivashoma, Mr</li>
+          <li>Nathoo Kusum J., Prof.</li>
+          <li>Fana Golden, Dr</li>
+          <li>Ncube Phumuzile, Ms</li>
+          <li>Gambanga Pauline, Dr</li>
+          <li>Ndamukwa Pikirai, Mr</li>
+          <li>Glavintcheva Iskra L., Dr</li>
+          <li>Ndhlovu Chiratidzo E., Prof</li>
+          <li>Gunguwo Hillary, Dr</li>
+          <li>Ndlovu Misheck, Mr</li>
+          <li>Gwanzura Lovemore, Prof.</li>
+          <li>Ndowa Francis, Dr</li>
+          <li>Gwata Beatrice, Mrs</li>
+          <li>Nembiri Tinashe,</li>
+          <li>Ms Hove Ropafadzai Mrs</li>
+          <li>Ngwende Gift W., Dr.</li>
+          <li>Kambarani Rose, Prof</li>
+          <li>Nkala Lee, Mr</li>
+          <li>Kandawasvika Petronella, Dr</li>
+          <li>Nyadzayo Tasiana K., Mr</li>
+          <li>Khoza Star, Dr</li>
+          <li>Nyakabau Anna M., Dr</li>
+          <li>Khumalo Brian, Mr</li>
+          <li>Nyamayaro Raphael, Dr</li>
+          <li>Khumalo Mhlawempi, </li>
+          <li>Nyaruwanga Albert, Mr</li>
+          <li>Kufa Tarisai, Dr</li>
+          <li>Pasi Christopher, Dr</li>
+          <li>Kusemwa Muyambi Preetyosa, Ms</li>
+          <li>Phiri Isaac, Dr</li>
+          <li>Latif AS, Prof</li>
+          <li>Reid Andrew, Dr</li>
+          <li>Machisa Vimbainashe, Ms</li>
+          <li>Rimai Ruth, Ms</li>
+          <li>Madhombiro Munyaradzi, Dr</li>
+          <li>Samukange Emma, Mrs</li>
+          <li>Madzikwa Newman B. Mr</li>
+          <li>Sandy Charles, Dr</li>
+          <li>Magombeyi Rudo, Ms</li>
+          <li>Sanyanga Arthur, Mr</li>
+          <li>Magunda Farai, </li>
+          <li>Shumba Godfrey, Mr</li>
+          <li>Mandimika, Florence</li>
+          <li>Sibanda Elopy N., Prof</li>
+          <li>Mandire Joice, Ms</li>
+          <li>Sifeku Florah N., Mrs</li>
+          <li>Mangezi Walter, Dr</li>
+          <li>Sithole Dorcas, Ms</li>
+          <li>Mangoma Tariro, Ms</li>
+          <li>Tagwirei Dexter, Prof</li>
+          <li>Mangwiro John C., Dr</li>
+          <li>Takaruza Kelvin, Mr</li>
+          <li>Masanganise Rangarirai, Prof</li>
+          <li>Tambudze Gaundencia, Ms</li>
+          <li>Masendu Maureen, Dr</li>
+          <li>Tekasala Lumbu Jerry, Mr</li>
+          <li>Mashinge Farayi, Mr</li>
+          <li>Ticklay Ismail, Dr</li>
+          <li>Mashoko Tsungai, Ms</li>
+          <li>Vuragu Davison N., Mr</li>
+          <li>Mashumba Azza, Dr</li>
+          <li>Wellington Maureen, Dr</li>
+          <li>Maswaure Laucas,</li>
+          <li>Zaranyika Trust, Dr</li>
+          <li>Matonhodze Alex, Mr</li>
+        </ul>
+        <p><strong>Thank You!</strong></p>
+        <p>Mrs R.F. Hove <br /><small>Director of Pharmacy Services </small></p>
+        <p>Prof. C. E. Ndhlovu<br /><small>NMTPAC Chairperson</small></p>
+      </span>
+};
