+++ src/Data/Information/foreword.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+export let foreword = {
+    short_title: "Foreword",
+    long_title: "Foreword on the 7th Edition",
+    content:
+      <span>
+        <h1>Foreword on the 7th Edition</h1>
+        <p>It is the national objective that the health care needs of Zimbabweans are met through the provision and proper use of essential medicines. Sometimes we do not need to give medicines, that is, there is not always a - "pill for every ill".  Thus, there is need to use medicines appropriately, efficiently, and effectively.</p>
+        <p>The guidelines in EDLIZ have always reflected the consensus of local experts, and takes into consideration factors such as the Zimbabwean setting, prevailing economic climate, practical experience as well as evidence-based therapeutics.</p>
+        <p>This new EDLIZ has taken into account the dynamic changes in the Burden of Disease as reflected by the inclusion of antiretroviral medicines and treatment of other opportunistic infections other than Tuberculosis (TB). Many of the therapeutic regimens of the previous EDLIZ still hold true and remain the same, and should reinforce the confidence of the prescriber in making reliable therapeutic choices.</p>
+        <p>I urge all health workers to familiarise themselves with the revised guidelines, to prescribe within the bounds of this publication, and to recognise the critical importance of providing a quality service to all health care recipients through the rational use of medicines.</p>
+        <p>EDLIZ <strong>REMAINS</strong> good medicine! Use it.</p>
+        <hr />
+        <p>Hon. Dr. P.D, Parirenyatwa<br /><small> Minister of Health &amp; Child Care </small></p>
+      </span>
+  }
