--- src/Data/Information/major_highlights.tsx
@@ -1,51 +1,0 @@
-import React from 'react'
-
-export let major_highlights = {
-    short_title: 'Highlights ',
-    long_title: 'Major Highlights in the Latest EDLIZ',
-    content: (
-        <span>
-            <h1>Major Highlights in the Latest EDLIZ</h1>
-            <h2>Preamble</h2>
-            <p>
-                The major changes in this latest edition of EDLIZ will be highlighted here so that you are aware of
-                recommendations that you need to consider in your medicine management or supply issues. Ideally each
-                hospital should create its own local medicine formulary which shows which medicines are considered very
-                useful in that setting so that you do not have to order medicines that your doctors will not prescribe
-                or use. For instance you should not keep specialist medicines if there is no specialist to prescribe
-                them. Hospital Medicine and Therapeutics Committees should select medicines for use in their hospital
-                using the EDLIZ.
-            </p>
-            <h2>New chapter</h2>
-            <p>There is one new chapter – Overview of Surgical Conditions.</p>
-            <h2>Antibiotics</h2>
-            <p>
-                A new cephalosporin, Cefixime, has been added to the treatment of sexually transmitted illnesses.
-                Azithromycin has also been added to the treatment of gonococcus.
-            </p>
-            <h2>Immunisation</h2>
-            <p>
-                Rotavirus immunisation has now become routinely available. Human Papilloma Virus (HPV) vaccination is
-                currently being used in a limited setting but it is hoped that it will be rolled out nationally in 2016.
-            </p>
-            <h2>Asthma Treatment</h2>
-            <p>
-                Given that salbutamol Inhalers are more accessible, <strong>oral</strong> salbutamol has been phased out
-                completely. You will need to ensure that your patients are aware of this change. Use of steroids as an
-                inhaler should be used in place of regular oral Salbutamol. Health care workers will need to check that
-                their clients can use the inhalers appropriately.
-            </p>
-            <h2>ART Guidelines (use latest ART guidelines)</h2>
-            <p>
-                Stavudine containing regimens are being phased out and will not be available except for a limited number
-                of patients who will still need them. For first line therapy, tenofovir/lamivudine and efavirenz will be
-                used in most instances for adults, adolescents and children as well as pregnant women.
-            </p>
-            <p>
-                Third line antiretroviral medicines will be available in selected hospitals. These third line medicines
-                will include raltegravirraltegravir and darunavir. Please familiarise yourself with the dosing of these
-                new medicines and the algorithm for their use.
-            </p>
-        </span>
-    ),
-}
