+++ src/Data/Information/preface.js
@@ -1,0 +1,112 @@
+import React from 'react';
+
+export let preface = {
+    short_title: "Preface",
+    long_title: "The Essential Medicines List for Zimbabwe",
+    content:
+      <span>
+        <h1>The Essential Medicines List for Zimbabwe - EDLIZ 7<sup>th</sup> Edition</h1>
+        <p>This 7th essential medicines list and standard treatment guidelines for the most common health conditions in Zimbabwe has been endorsed by the National Medicine & Therapeutics Policy Advisory Committee [NMTPAC]. It is the product of many years of combined efforts by hundreds of health workers at all levels of the health care system in Zimbabwe – from the front line health care providers to the providers of specialist care. It has been refined over the years as a result of its widespread use by our healthcare workers. We continue to revise the standard treatment guidelines and take into account medicine developments and new healthcare problems. Thus this latest edition has included more essential medicines.</p>
+        <p>The essential medicine list is based on the Essential Medicines Concept. Medicines in EDLIZ are chosen to meet the health care needs of the majority of the population, and should therefore always be available and accessible at a price that both the patient and the nation can afford.</p>
+        <h2>Selection of medicines for inclusion </h2>
+        <p>Selection of medicines for inclusion in EDLIZ has been based on the following criteria, with special emphasis on proven evidence for their use in the Zimbabwean setting:</p>
+        <ul>
+          <li>relevance to prevalent diseases</li>
+          <li>proven efficacy and safety </li>
+          <li>adequate scientific data in a variety of settings </li>
+          <li>adequate quality</li>
+          <li>favourable cost-benefit ratio</li>
+          <li>desirable pharmacokinetics</li>
+          <li>possibilities for local manufacture </li>
+          <li>available as single ingredient items </li>
+        </ul>
+        <ul className="boxedlist">
+          <li>Safe</li>
+          <li>Efficacious</li>
+          <li>Quality</li>
+          <li>Available</li>
+          <li>Affordable</li>
+          <li>Accessible</li>
+          <li>Rationally used</li>
+        </ul>
+        <h3>Generic Medicines</h3>
+        <p>Every medicine has a chemical name and a generic name. For example, paracetamol, its chemical name is N-(4-Hydroxyphenol) acetamide and the international non-proprietary name (INN) or generic name is paracetamol. The INN is the medicine's official name regardless of who manufactures or markets it. An additional brand name is chosen by the manufacturer to facilitate recognition and association of the product with a particular manufacturer for marketing purposes.</p>
+        <p>For most common medicines there are several branded products that all contain the same active ingredient and therefore share the same INN. For example, the <em>African Monthly Index of Medical Specialties</em> (MIMS) lists over fifteen different brand names of paracetamol. There are 12 different preparations containing aspirin, 13 different brands of amoxicillin and 12 different brands of ampicillin.</p>
+        <p>The use of generic names for medicine procurement as well as prescribing carries considerations of clarity, quality, and price. Proponents of generic medicines procurement and prescribing point out that:</p>
+        <ul>
+          <li>generic names are more informative than brand names and facilitate purchasing of products from multiple suppliers, whether as brandname or as generic products; </li>
+          <li>generic medicines are generally cheaper than products sold by brand name; this is demonstrateddemonstrated very clearly when it comes to antiretroviral medicines </li>
+          <li>generic prescribing also facilitates product substitution, whenever appropriate. </li>
+        </ul>
+        <p>Opponents argue that the quality of generic medicines is inferior to that of brand (innovator) products. However quality assurance and naming of medicines are completely separate issues. Generic medicines from reliable suppliers are as safe, effective, and high in quality as medicines with brand names. At the same time, branded medicines from a manufacturer with inadequate procedures for quality control can be of poor quality, despite the brand name. Also, although any medicine can be counterfeited, there are more incentives for counterfeiting brandname medicines than generic medicines. Some pharmaceutical companies also sell their branded products under the generic name, for a much lower price.</p>
+        <p>Bio-equivalence is often misused as an argument against the use of generic equivalents. For many medicines, the variation in bioavailability among individual patients is much larger than the variation among products of different manufacturers. In fact, bioavailability is clinically relevant for only a relatively small number of medicines such as furosemide, digoxin, levodopa, isoniazid, theophylline and phenytoin.</p>
+        <p>Zimbabwe has a well understood generic policy which requires that all prescribing is in the generic name and the dispenser can make generic substitutions (unless bioavailability is an issue in which case the prescriber should indicate accordingly).</p>
+        <h3>Advantages of EDLIZ</h3>
+        <p>The benefits of the selection and use of a limited number of essential medicines are:</p>
+        <ul>
+          <li>Improved medicines supply</li>
+          <li>More rational prescribing</li>
+          <li>Lower costs</li>
+          <li>Improved patient use</li>
+        </ul>
+        <h4>Improved Medicines Supply</h4>
+        <ul className="boxedlist">
+          <li>easier procurement, storage & distribution</li>
+          <li>lower holding stocks</li>
+          <li>lower losses</li>
+          <li>better quality assurance</li>
+        </ul>
+        <p>The regular supply of medicines is difficult in many countries, and the consequent health implications are many. Improved medicines availability should lead to improved clinical outcomes. With fewer essential medicines being purchased, the mechanisms and logistics for procurement, storage & distribution will clearly be easier. It is not practical for each clinic in Zimbabwe to attempt to procure, transport and warehouse all the hundreds of items in EDLIZ. Conversely, limiting the number of medicines available at the primary health care level makes a regular supply of medicines more practical and possible.</p>
+        <p>With an improved supply the possibilities of holding lower quantities exist. This has financial implications as well as reducing the likelihood of medicines expiring or being damaged during storage.</p>
+        <p>Quality assurance can be better managed when the number of medicines is limited, and quality checks can be performed more frequently.</p>
+        <h3>More Rational Prescribing</h3>
+        <ul className="boxedlist">
+          <li>focused, more effective training </li>
+          <li>more experience with fewer medicines</li>
+          <li>no irrational treatment alternatives available</li>
+          <li>focused medicine information </li>
+          <li> better recognition of adverse medicine reactions</li>
+        </ul>
+        <p>In the absence of limited lists, the large variety of products available on the market contributes to inconsistent prescribing and consequently, variation in clinical practice even within the same health care facility. Irrational prescribing may lead to therapeutic hazards and increased costs.</p>
+        <p>When the number of medicines is limited, training can be more focused and the quality of care enhanced. This is especially true when the list represents a consensus of opinion on first choice of treatment such as in EDLIZ</p>
+        <p>Using EDLIZ enables the prescriber to become more familiar with the medicines they use, and better able to recognise adverse effects.</p>
+        <p>The use of EDLIZ also eliminates irrational products from being available for prescribing, and allows for more focused medicine information to be provided on suitable essential medicines</p>
+        <h3>Lower Costs</h3>
+        <ul className="boxedlist">
+          <li>more competition</li>
+          <li>lower prices</li>
+        </ul>
+        <p>Improved effectiveness and efficiency in patient treatment leads to lower health care costs. The essential medicines concept is increasingly being accepted as a universal tool to promote both quality of care and cost control.</p>
+        <p>Essential medicines are usually available from multiple suppliers. With increased competition, more favourable prices can be negotiated.</p>
+        <p>By limiting the number of different medicines that can be used to treat a particular clinical problem, larger quantities of the selected medicine will be needed, with potential opportunities to achieve economies of scale.</p>
+        <h3>Improved Patients Use</h3>
+        <p>Focusing on fewer medicines can enhance patient education and efforts to promote the proper use of medicines in both patients and prescribers</p>
+        <ul className="boxedlist">
+          <li>focused education efforts</li>
+          <li>reduced confusion & increased adherence to treatment </li>
+        </ul>
+        <p>Additionally, with improved medicine availability changes to chronic medication regimens are less likely and as a consequence patients have a better understanding of their disease, their medication and the need for compliance.</p>
+        <h2>Implementation of EDLIZ and Setting up of Hospital Medicine and Therapeutics Committees (HMTCS)</h2>
+        <p>The advantages presented here however do not just happen. EDLIZ itself will not ensure rational prescribing or facilitate good procurement or quality assurance. Educational, regulatory, financial or managerial strategies on their own are less effective in promoting the rational use of medicines than combined strategies. The production of EDLIZ is one such regulatory strategy, but further steps such as training and retraining, patient education and the establishment and effective functioning of hospital medicine and therapeutic committees (HMTCs) have to be taken to ensure cost-effective prescribing and patient care. It is therefore necessary for every hospital to have a forum where medicine issues can be discussed. Ideally, a separate hospital medicine and therapeutics committee (HMTC) should be formed. Given the current manpower constraints, we encourage hospitals to exploit every opportunity such as the regular divisionaldivisional meetings heldheld in Central Hospitals to discuss and address medicine related problems.   The NMTPAC is available to assist those hospitals that are ready to set up an HMTC. A technical guideline to set up a HMTC developed by the NMTPAC is available.</p>
+        <h2>Explanations and Changes from the Previous Version </h2>
+        <p>This edition is essentially the same in format and layout, categorisation as the last edition.  You will need to read it carefully to note changes in recommendations that apply to your areas of interest. Extra bulletins will be sent out where drastic changes in medicine recommendations have occurred.</p>
+        <p>All medicines in EDLIZ are categorised firstly by level of availability (ABCS) in the health care system, and secondly, according to priority (VEN). Hence in the example below, amoxicillin is available at primary health care facility (C) level and is ranked vital (V).</p>
+        <table>
+          <tbody>
+            <tr><th>Medicine</th><th>Codes</th><th>Adult Dose</th><th>Frequency</th><th>Duration</th></tr>
+            <tr><td>amoxicillin po</td><td>C V</td><td>500mg</td><td>3 times a day</td><td>7 Days</td></tr>
+          </tbody>
+        </table>
+        <h3>Level of Availability</h3>
+        <p><strong>C medicines</strong> are those required at primary health care level and should be available at all levels of care</p>
+        <p><strong>B medicines</strong> are found at district hospital level or secondary and higher levels of care. Some B medicines may be held at primary health care</p>
+        <p><strong>A medicines</strong> are prescribed at provincial or central hospital levels</p>
+        <p><strong>S medicines</strong> (specialist only) have been brought back into this edition. These are medicines that require special expertise and /or diagnostic tests before being prescribed.</p>
+        <h2>VEN Classification</h2>
+        <p>All medicines are also classified according to their priority. This is mostly a tool to assist in giving priority to medicines based on economic considerations. Thus <strong>V</strong> medicines are <strong>vital</strong>, they are considered lifesaving or their unavailability would cause serious harm and efforts should always be aimed at making them 100% available.</p>
+        <p><strong>E</strong> medicines are <strong>essential</strong>, and are given second priority. Without E medicines there would be major discomfort or irreversible harm. And <strong>N</strong> medicines are still <strong>necessary</strong> but are lower in priority than V and E medicines.</p>
+        <p>The NMTPAC is a standing committee that reviews the therapeutic guidelines in EDLIZ on a continual basis, and always looks forward to feedback from the providers of health care in Zimbabwe.  Contact the NMTPAC through Directorate of Pharmacy Services on <a href="mailto:dps@mohcc.gov.zw">dps@mohcc.gov.zw</a> or <a href="mailto:nmtpacnmtpac@gmail.com">nmtpacnmtpac@gmail.com</a>  with your comments.</p>
+        <p><small>
+          Brigadier General (Dr) G. Gwinji <br />Permanent Secretary <br />Ministry of Health & Child Care <br />Republic of Zimbabwe </small></p>
+      </span>
+  }
