+++ src/Data/Information/seventh_edition.js
@@ -1,0 +1,41 @@
+import React from 'react';
+
+export let seventh_edition = {
+    short_title: "7th Edition",
+    long_title: "EDLIZ 7th Edition 2015",
+    content:
+      <span>
+        <h1>Published By:</h1>
+        <p>The National Medicine and Therapeutics Policy Advisory Committee (NMTPAC) <br /> Ministry of Health & Child Care Republic of Zimbabwe</p>
+        <p>Further copies may be obtained through the relevant Provincial Medical Directorate, City Health Directorate, the NMTPAC, Ministry of Health & Child Care (MOHCC), PO Box CY 1122, Causeway, Harare, Zimbabwe, or the <a href="http://www.mohcc.gov.zw" target="_blank" rel="noopener noreferrer">MOHCC website</a>. Copies of the text may be obtained on soft copy if required for teaching purposes from email address: <a href="mailto:dps@mohcc.gov.zw">dps@mohcc.gov.zw</a> or <a href="mailto:nmtpac@gmail.com">nmtpac@gmail.com</a> EDLIZ was prepared using Microsoft® Word.  The information presented in these guidelines conforms to current medical, nursing and pharmaceutical practice. It is provided in good faith. Whilst every effort was made to ensure that medicine doses are correct, no responsibility can be taken for errors and omissions.</p>
+        <p><strong>- EDLIZ Review Co-ordinator </strong></p>
+        <p>No part of this publication may be reproduced by any process without the written permission of the copyright holder, exception being made for the purpose of private study, research, criticism or review, or for teaching, but not for sale or other commercial use.</p>
+        <p>Original Cover Design: Regina Gapa and Charon Lessing.</p>
+        <p>Cover redesign & Layout: Kim Hoppenworth.</p>
+        <p>Cover redesign and layout 2015: Newman B Madzikwa</p>
+        <p>&copy; Copyright June 2015, Ministry of Health &amp; Child Care</p>
+        <h1>EDLIZ Review Committee</h1>
+        <ul>
+          <li>Apollo Tsitsi</li>
+          <li>Samukange Emma</li>
+          <li>Borok Margaret</li>
+          <li>Mujuru Hilda</li>
+          <li>Chakanyuka Christine C.</li>
+          <li>Mushavi Angela</li>
+          <li>Hove Ropafadzai</li>
+          <li>Madzikwa Newman B.</li>
+          <li>Mungwadzi Godfrey</li>
+          <li>Ndhlovu Chiratidzo E.</li>
+          <li>Bakasa Clemenciana</li>
+          <li>Sifeku Florah N.</li>
+          <li>Mudzimu Forward</li>
+          <li>Wellington Maureen</li>
+          <li>Torongo Mabel</li>
+          <li>Vuragu Davison N.</li>
+          <li>Nyamayaro Raphael</li>
+          <li>Maunganidze Aspect</li>
+          <li>Khoza Star</li>
+          <li> Basopo Victor</li>
+        </ul>
+      </span>
+  }
