+++ src/Views/HamburgerButton/hamburgerbutton.css
@@ -1,3 +1,20 @@
-.night > .hamburger-button {
-    background-color: #2a2a2a;
+.hamburger-button{
+    position:fixed; 
+    top: 0; 
+    left:0; 
+    border: none;
+    z-index: 99999; 
+    padding: 16px; 
+    margin:0; 
+    box-sizing:"border-box"; 
+    background-color: #fff;
+    width: 50px; 
+    height: 50px;
+    outline: none;
+    border-radius: 50%;
 }
+
+
+.night > .hamburger-button {
+  background-color: #2a2a2a;
+}
