+++ src/Views/NavigationItem/navigationitem.css
@@ -1,11 +1,32 @@
-.nav-active .nav-long-title {
-    color: #2b44ff;
+.nav-item {
+    display: block;
+    padding: 10px;
+    text-align:left;
+    border-bottom:1px solid rgba(200,200,200, 0.2);
+    text-decoration: none;
 }
 
-.night .nav-short-title {
+.nav-active{
+    border-left:3px solid #6c63ff;
+    padding-left: 7px;
+}
+
+
+.nav-short-title{
+    color: #999;
+    display: block;
+    font-size:0.8em;
+}
+.night .nav-short-title{
     color: #ddd;
 }
 
-.night .nav-long-title {
+.nav-long-title{
+    display: block;
+    color:#6c63ff;
+    margin:5px 0;
+}
+
+.night .nav-long-title{
     color: #fff;
 }
