+++ src/Views/ReaderScreen/readerscreen.css
@@ -1,3 +1,17 @@
+.screen {
+    position: absolute;
+    top: 0;
+    bottom: 50px;
+    width: 100vw;
+    box-sizing: border-box;
+    right: 0;
+    left: 0;
+    padding: 0em;
+    text-align: left;
+    height: calc(100vh);
+    overflow-y: scroll;
+}
+
 .day .screen {
     background-color: #fff;
     color: #333;
@@ -8,12 +22,21 @@
     color: #ccc;
 }
 
-p {
-    font-family: 'IBM Plex Serif', serif;
-    font-size: 18px;
-    color: #101010;
-    letter-spacing: 0.26px;
-    line-height: 30px;
+.screen::-webkit-scrollbar-track {
+    box-shadow: inset 0 0 1px rgba(0, 0, 0, 0.2);
+    border-radius: 2px;
+    background-color: rgba(0, 0, 0, 0);
+}
+
+.screen::-webkit-scrollbar {
+    width: 3px;
+    background-color: rgba(0, 0, 0, .1);
+}
+
+.screen::-webkit-scrollbar-thumb {
+    border-radius: 3px;
+    box-shadow: inset 0 0 11px rgba(0, 0, 0, .2);
+    background-color: #6c63ff;
 }
 
 ul,
@@ -24,10 +47,8 @@ ol {
 li {
     margin-bottom: 0.5em;
 }
-h1 {
-    display: none;
-}
 
+h1,
 h2,
 h3,
 h4,
@@ -36,9 +57,27 @@ h6 {
     text-align: left;
 }
 
+.article {
+    display: block;
+    margin: auto;
+    padding: 60px 1em 60px;
+    max-width: 840px;
+}
+
+.screen a {
+    border-bottom: 1px dotted;
+}
+
 .night .screen a {
     color: #ccc;
 }
+
+/* a {
+    text-decoration: none;
+    color: inherit;
+    display: inline-block;
+    border-bottom: 1px dotted;
+} */
 .underline {
     text-decoration: underline;
 }
@@ -66,6 +105,8 @@ table {
     background-color: rgba(255, 255, 255, 0.2);
 }
 
+
+
 @media screen and (max-width: 450px) {
     table {
         font-size: 85%;
@@ -80,8 +121,8 @@ table {
 
     .screen::-webkit-scrollbar-thumb {
         border-radius: 0;
-        -webkit-box-shadow: inset 0 0 11px rgba(0, 0, 0, 0.3);
-        box-shadow: inset 0 0 11px rgba(0, 0, 0, 0.3);
+        -webkit-box-shadow: inset 0 0 11px rgba(0, 0, 0, .3);
+        box-shadow: inset 0 0 11px rgba(0, 0, 0, .3);
         background-color: #ccf;
     }
 }
