--- src/index.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
-        <meta name="theme-color" content="#0093e9" />
-        <meta name="mobile-web-app-capable" content="yes" />
-        <link rel="manifest" href="../public/manifest.json" />
-        <link rel="shortcut icon" href="../public/android/android-launchericon-96-96.png" />
-        <!-- Add to home screen for Safari on iOS -->
-        <meta name="apple-mobile-web-app-capable" content="yes" />
-        <meta name="apple-mobile-web-app-status-bar-style" content="blue" />
-        <meta name="apple-mobile-web-app-title" content="EDLIZ" />
-        <link rel="apple-touch-icon" href="../public/ios/ios-appicon-180-180.png" />
-        <!-- Tile icons on Windows -->
-        <meta name="msapplication-TileImage" content="../public/windows/windowsphone-mediumtile-150-150.png" />
-        <meta name="msapplication-TileColor" content="#0093e9" />
-        <!--Twitter Card meta-data-->
-        <meta name="twitter:card" content="summary_large_image" />
-        <meta name="twitter:site" content="@michaeldera" />
-        <meta name="twitter:creator" content="@michaeldera" />
-        <meta name="twitter:title" content="EDLIZ | Essential Drugs and Medicines List  for Zimbabwe " />
-        <meta
-            name="twitter:description"
-            content="This 7th essential medicines list and standard treatment guidelines for the most common health conditions in Zimbabwe has been endorsed by the National Medicine & Therapeutics Policy Advisory Committee [NMTPAC]."
-        />
-        <meta name="twitter:image" content="https://edliz.azurewebsites.net/img/pills.jpg" />
-        <!--Facebook Opengraph tags-->
-        <meta property="og:url" content="https://edliz.azurewebsites.net/" />
-        <meta property="og:type" content="website" />
-        <meta property="og:title" content="EDLIZ  | Essential Drugs and Medicines List  for Zimbabwe" />
-        <meta
-            property="og:description"
-            content="This 7th essential medicines list and standard treatment guidelines for the most common health conditions in Zimbabwe has been endorsed by the National Medicine & Therapeutics Policy Advisory Committee [NMTPAC]."
-        />
-        <meta property="og:image" content="https://edliz.azurewebsites.net/img/pills.jpg" />
-        <link href="https://fonts.googleapis.com/css?family=IBM+Plex+Serif|Muli:600" rel="stylesheet" />
-        <script src="../public/js/modernizr.custom.js" content-type="text/javascript"></script>
-        <title>Essential Drugs and Medicines List in Zimbabwe</title>
-        <title>EDLIZ | Essential Drugs and Medicines List for Zimbabwe</title>
-    </head>
-    <body>
-        <div id="app" />
-    </body>
-</html>
